Lab10 polyadd.py
Programming Lab 10: polyadd.py
Think of that you are given two polynomial, say f(x) = x
2 − 4x + 6 and g(x) = 2x
3 + x.
Adding two polynomials is easier than multiplying. For example, adding them results in
f(x) + g(x) = 2x
3 + x
2 − 3x + 6.
Given two polynomials, compute the addition result of them. The coefficients an the
polynomial f(x) = anx
n + an−1x
n−1 + · · · + a1x + a0 can be given as input. Note that you have to consider the coefficients though they are zeros. For instance, the polynomial g(x) = 2x^3 has four coefficients which are 2, 0, 0, and 0.

In the first line of the standard input, the coefficients of it are given separated by a space; in the second line, those of the second. Your program should print the sum of the polynomial in a pretty style. In the pretty style, the exponent should be printed after the caret, say, x ^ 2 should be printed in x^2. The pretty style should not contain the term with zero coefficients, should not print coefficient one except the constant term, also the negative coefficient should be printed as a subtracted term except for that of the term with the highest degree, where
the negative coefficient should be printed as is. There should be no spaces between the coefficients and the variable x, but the terms should be separated by the addition or the subtraction operator with spaces on both sides.

Additional requirements for bonus points
1. You have to define and use the class representing a polynomial.
2. You have to overload the add operator (+).


Input Output
1 -4 6
2 0 1 0
2x^3 + x^2 - 3x + 6
-2 0 0 0
1 1
-2x^3 + x + 1